#421 refactor(workspace): reuse public markdown review surface
